revert type and add pictureURL column to the background table

+++ src/main/java/com/StudyMate/StudyMate/model/Favourite_background.java
@@ -1,22 +1,22 @@
 package com.StudyMate.StudyMate.model;
 
 public class Favourite_background {
-    private Long userID;
+    private Integer userID;
     private String backgroundID;
 
     public Favourite_background() {
     }
 
-    public Favourite_background(Long userID, String backgroundID) {
+    public Favourite_background(Integer userID, String backgroundID) {
         this.userID = userID;
         this.backgroundID = backgroundID;
     }
 
-    public Long getUserID() {
+    public Integer getUserID() {
         return userID;
     }
 
-    public void setUserID(Long userID) {
+    public void setUserID(Integer userID) {
         this.userID = userID;
     }
